+++ app/src/main/java/com/example/playlistmaker/di/RepositoryModule.kt
@@ -1,9 +1,9 @@
 package com.example.playlistmaker.di
 
-import com.example.playlistmaker.player.data.repository.PlayerGetTrackRepositoryImpl
-import com.example.playlistmaker.player.data.repository.mediaplayer.MediaPlayerRepositoryImpl
-import com.example.playlistmaker.player.domain.repository.MediaPlayerRepository
-import com.example.playlistmaker.player.domain.repository.PlayerGetTrackRepository
+import com.example.playlistmaker.media.player.data.repository.PlayerGetTrackRepositoryImpl
+import com.example.playlistmaker.media.player.data.repository.mediaplayer.MediaPlayerRepositoryImpl
+import com.example.playlistmaker.media.player.domain.repository.MediaPlayerRepository
+import com.example.playlistmaker.media.player.domain.repository.PlayerGetTrackRepository
 import com.example.playlistmaker.search.data.repository.FindTrackRepositoryImpl
 import com.example.playlistmaker.search.data.repository.TrackHistoryRepositoryImpl
 import com.example.playlistmaker.search.domain.repository.FindTrackRepository
@@ -25,7 +25,7 @@ val repositoryModule = module {
 
 //PLAYER
     single<PlayerGetTrackRepository> { PlayerGetTrackRepositoryImpl() }
-    factory<MediaPlayerRepository> {MediaPlayerRepositoryImpl(get())  }
+    factory<MediaPlayerRepository> { MediaPlayerRepositoryImpl(get())  }
 
     //SETTINGS
     single<ThemeChangeRepository> { ThemeChangeRepositoryImpl(get()) }
