+++ app/src/main/java/com/example/playlistmaker/media/player/domain/repository/PlayerGetTrackRepository.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.media.player.domain.repository
+
+import com.example.playlistmaker.media.player.ui.models.PlayerTrack
+
+interface PlayerGetTrackRepository {
+    fun getPlayerTrack(trackGson:String?): PlayerTrack
+}
