--- app/src/main/java/com/example/playlistmaker/player/domain/repository/PlayerGetTrackRepository.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.player.domain.repository
-
-import com.example.playlistmaker.player.ui.models.PlayerTrack
-
-interface PlayerGetTrackRepository {
-    fun getPlayerTrack(trackGson:String?): PlayerTrack
-}
