--- app/src/main/java/com/example/playlistmaker/player/domain/usecases/GetPlayerTrackUseCase.kt
@@ -1,11 +1,0 @@
-package com.example.playlistmaker.player.domain.usecases
-import com.example.playlistmaker.player.ui.models.PlayerTrack
-import com.example.playlistmaker.player.domain.repository.PlayerGetTrackRepository
-
-
-class GetPlayerTrackUseCase (private val playerTrackRepository: PlayerGetTrackRepository) {
-    fun execute(trackGson:String?): PlayerTrack {
-        return playerTrackRepository.getPlayerTrack(trackGson)
-    }
-
-}
