+++ app/src/main/res/drawable/empty_hearct_icon.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="25dp"
+    android:height="23dp"
+    android:viewportWidth="25"
+    android:viewportHeight="23">
+  <path
+      android:pathData="M18.125,0C15.95,0 13.863,1.015 12.5,2.62C11.137,1.015 9.05,0 6.875,0C3.025,0 0,3.033 0,6.894C0,11.632 4.25,15.492 10.688,21.358L12.5,23L14.313,21.346C20.75,15.492 25,11.632 25,6.894C25,3.033 21.975,0 18.125,0ZM12.625,19.49L12.5,19.616L12.375,19.49C6.425,14.088 2.5,10.516 2.5,6.894C2.5,4.387 4.375,2.507 6.875,2.507C8.8,2.507 10.675,3.748 11.337,5.465H13.675C14.325,3.748 16.2,2.507 18.125,2.507C20.625,2.507 22.5,4.387 22.5,6.894C22.5,10.516 18.575,14.088 12.625,19.49Z"
+      android:fillColor="#ffffff"/>
+</vector>
