+++ app/src/main/res/drawable/player_button_shape_light.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+    <solid android:color="@color/YP_lightGray_button" />
+</shape>
